refactored Cache and rename to LineNumberCache
